feat(favorite): move FavoriteRepository to domain layer
