+++ shared/data/movies/src/main/java/com/my/movie/favorite/di/FavoriteFactory.kt
@@ -1,9 +1,9 @@
 package com.my.movie.favorite.di
 
 import android.content.Context
-import com.my.movie.favorite.FavoriteRepository
-import com.my.movie.favorite.storage.FavoriteDatabase
+import com.my.domain.repository.FavoriteRepository
 import com.my.movie.favorite.FavoriteRepositoryImpl
+import com.my.movie.favorite.storage.FavoriteDatabase
 
 class FavoriteFactory {
     fun provideRepository(context: Context): FavoriteRepository {
